Fixed integer division instead of float division when setting health bars

+++ src/com/ahsgaming/valleyofbones/screens/panels/InfoPanel.java
@@ -95,7 +95,7 @@ public class InfoPanel extends Panel {
             }
 
             healthBar.setSize(lblHealth.getWidth(), 4);
-            healthBar.setCurrent(selected.getCurHP() / selected.getMaxHP());
+            healthBar.setCurrent((float)selected.getCurHP() / selected.getMaxHP());
         }
 
         dirty = true;
